+++ sample/proto_v2/role_cfg.amd.js
@@ -107,8 +107,8 @@ define({
           id : 501
         }
       },
-      cost_value : 99999,
-      id : 10005
+      id : 10005,
+      cost_value : 99999
     },
     {
       int_as_string : "6",