+++ sample/proto_v3/role_cfg.lua
@@ -107,8 +107,8 @@ return {
                     id = 501,
                 },
             },
-            cost_value = 99999,
             id = 10005,
+            cost_value = 99999,
         },
         {
             int_as_string = "6",