+++ src/org/xresloader/core/data/dst/DataDstPbHelper.java
@@ -7,6 +7,7 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.UnsupportedEncodingException;
 import java.nio.charset.Charset;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
@@ -14,12 +15,21 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
+import java.util.List;
 import java.util.Map;
+import java.util.Set;
+
 import com.google.protobuf.ByteString;
 import com.google.protobuf.DescriptorProtos;
 import com.google.protobuf.Descriptors;
+import com.google.protobuf.Descriptors.EnumValueDescriptor;
+import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.Descriptors.FieldDescriptor.JavaType;
+import com.google.protobuf.ExtensionRegistry.ExtensionInfo;
 import com.google.protobuf.DynamicMessage;
+import com.google.protobuf.GeneratedMessageV3;
+import com.google.protobuf.InvalidProtocolBufferException;
+import com.google.protobuf.Message;
 import com.google.protobuf.UninitializedMessageException;
 import org.apache.commons.codec.binary.Hex;
 import org.xresloader.Xresloader;
@@ -149,19 +159,16 @@ public class DataDstPbHelper {
 
     // ==================== dump options ====================
 
-    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.EnumValueDescriptor msg_desc) {
+    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.EnumValueDescriptor msg_desc,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
         HashMap<String, Object> msg_root = new HashMap<String, Object>();
         boolean has_data = false;
 
-        Map<Descriptors.FieldDescriptor, Object> options_data = msg_desc.getOptions().getAllFields();
-        if (!options_data.isEmpty()) {
-            HashMap<String, Object> options_root = new HashMap<String, Object>();
+        HashMap<String, Object> options_data = dumpMessageExtensions(msg_desc.getOptions(),
+                msg_desc.getOptions().getAllFields(), custom_extensions);
+        if (options_data != null && !options_data.isEmpty()) {
             has_data = true;
-            msg_root.put("options", options_root);
-
-            for (Map.Entry<Descriptors.FieldDescriptor, Object> data : options_data.entrySet()) {
-                options_root.put(data.getKey().getName(), data.getValue());
-            }
+            msg_root.put("options", options_data);
         }
 
         if (has_data) {
@@ -173,25 +180,22 @@ public class DataDstPbHelper {
         return null;
     }
 
-    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.EnumDescriptor msg_desc) {
+    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.EnumDescriptor msg_desc,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
         HashMap<String, Object> msg_root = new HashMap<String, Object>();
         boolean has_data = false;
 
-        Map<Descriptors.FieldDescriptor, Object> options_data = msg_desc.getOptions().getAllFields();
-        if (!options_data.isEmpty()) {
-            HashMap<String, Object> options_root = new HashMap<String, Object>();
+        HashMap<String, Object> options_data = dumpMessageExtensions(msg_desc.getOptions(),
+                msg_desc.getOptions().getAllFields(), custom_extensions);
+        if (options_data != null && !options_data.isEmpty()) {
             has_data = true;
-            msg_root.put("options", options_root);
-
-            for (Map.Entry<Descriptors.FieldDescriptor, Object> data : options_data.entrySet()) {
-                options_root.put(data.getKey().getName(), data.getValue());
-            }
+            msg_root.put("options", options_data);
         }
 
         // value
         HashMap<String, Object> child_message = null;
         for (Descriptors.EnumValueDescriptor sub_desc : msg_desc.getValues()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -213,19 +217,16 @@ public class DataDstPbHelper {
         return null;
     }
 
-    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.OneofDescriptor msg_desc) {
+    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.OneofDescriptor msg_desc,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
         HashMap<String, Object> msg_root = new HashMap<String, Object>();
         boolean has_data = false;
 
-        Map<Descriptors.FieldDescriptor, Object> options_data = msg_desc.getOptions().getAllFields();
-        if (!options_data.isEmpty()) {
-            HashMap<String, Object> options_root = new HashMap<String, Object>();
+        HashMap<String, Object> options_data = dumpMessageExtensions(msg_desc.getOptions(),
+                msg_desc.getOptions().getAllFields(), custom_extensions);
+        if (options_data != null && !options_data.isEmpty()) {
             has_data = true;
-            msg_root.put("options", options_root);
-
-            for (Map.Entry<Descriptors.FieldDescriptor, Object> data : options_data.entrySet()) {
-                options_root.put(data.getKey().getName(), data.getValue());
-            }
+            msg_root.put("options", options_data);
         }
 
         if (has_data) {
@@ -236,19 +237,16 @@ public class DataDstPbHelper {
         return null;
     }
 
-    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.FieldDescriptor msg_desc) {
+    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.FieldDescriptor msg_desc,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
         HashMap<String, Object> msg_root = new HashMap<String, Object>();
         boolean has_data = false;
 
-        Map<Descriptors.FieldDescriptor, Object> options_data = msg_desc.getOptions().getAllFields();
-        if (!options_data.isEmpty()) {
-            HashMap<String, Object> options_root = new HashMap<String, Object>();
+        HashMap<String, Object> options_data = dumpMessageExtensions(msg_desc.getOptions(),
+                msg_desc.getOptions().getAllFields(), custom_extensions);
+        if (options_data != null && !options_data.isEmpty()) {
             has_data = true;
-            msg_root.put("options", options_root);
-
-            for (Map.Entry<Descriptors.FieldDescriptor, Object> data : options_data.entrySet()) {
-                options_root.put(data.getKey().getName(), data.getValue());
-            }
+            msg_root.put("options", options_data);
         }
 
         if (has_data) {
@@ -261,25 +259,22 @@ public class DataDstPbHelper {
         return null;
     }
 
-    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.Descriptor msg_desc) {
+    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.Descriptor msg_desc,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
         HashMap<String, Object> msg_root = new HashMap<String, Object>();
         boolean has_data = false;
 
-        Map<Descriptors.FieldDescriptor, Object> options_data = msg_desc.getOptions().getAllFields();
-        if (!options_data.isEmpty()) {
-            HashMap<String, Object> options_root = new HashMap<String, Object>();
+        HashMap<String, Object> options_data = dumpMessageExtensions(msg_desc.getOptions(),
+                msg_desc.getOptions().getAllFields(), custom_extensions);
+        if (options_data != null && !options_data.isEmpty()) {
             has_data = true;
-            msg_root.put("options", options_root);
-
-            for (Map.Entry<Descriptors.FieldDescriptor, Object> data : options_data.entrySet()) {
-                options_root.put(data.getKey().getName(), data.getValue());
-            }
+            msg_root.put("options", options_data);
         }
 
         // nested_type
         HashMap<String, Object> child_message = null;
         for (Descriptors.Descriptor sub_desc : msg_desc.getNestedTypes()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -296,7 +291,7 @@ public class DataDstPbHelper {
         // enum_type
         child_message = null;
         for (Descriptors.EnumDescriptor sub_desc : msg_desc.getEnumTypes()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -313,7 +308,7 @@ public class DataDstPbHelper {
         // oneof_decl
         child_message = null;
         for (Descriptors.OneofDescriptor sub_desc : msg_desc.getOneofs()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -330,7 +325,7 @@ public class DataDstPbHelper {
         // field
         child_message = null;
         for (Descriptors.FieldDescriptor sub_desc : msg_desc.getFields()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -352,19 +347,16 @@ public class DataDstPbHelper {
         return null;
     }
 
-    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.MethodDescriptor msg_desc) {
+    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.MethodDescriptor msg_desc,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
         HashMap<String, Object> msg_root = new HashMap<String, Object>();
         boolean has_data = false;
 
-        Map<Descriptors.FieldDescriptor, Object> options_data = msg_desc.getOptions().getAllFields();
-        if (!options_data.isEmpty()) {
-            HashMap<String, Object> options_root = new HashMap<String, Object>();
+        HashMap<String, Object> options_data = dumpMessageExtensions(msg_desc.getOptions(),
+                msg_desc.getOptions().getAllFields(), custom_extensions);
+        if (options_data != null && !options_data.isEmpty()) {
             has_data = true;
-            msg_root.put("options", options_root);
-
-            for (Map.Entry<Descriptors.FieldDescriptor, Object> data : options_data.entrySet()) {
-                options_root.put(data.getKey().getName(), data.getValue());
-            }
+            msg_root.put("options", options_data);
         }
 
         if (has_data) {
@@ -377,25 +369,22 @@ public class DataDstPbHelper {
         return null;
     }
 
-    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.ServiceDescriptor msg_desc) {
+    static public HashMap<String, Object> dumpOptionsIntoHashMap(Descriptors.ServiceDescriptor msg_desc,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
         HashMap<String, Object> msg_root = new HashMap<String, Object>();
         boolean has_data = false;
 
-        Map<Descriptors.FieldDescriptor, Object> options_data = msg_desc.getOptions().getAllFields();
-        if (!options_data.isEmpty()) {
-            HashMap<String, Object> options_root = new HashMap<String, Object>();
+        HashMap<String, Object> options_data = dumpMessageExtensions(msg_desc.getOptions(),
+                msg_desc.getOptions().getAllFields(), custom_extensions);
+        if (options_data != null && !options_data.isEmpty()) {
             has_data = true;
-            msg_root.put("options", options_root);
-
-            for (Map.Entry<Descriptors.FieldDescriptor, Object> data : options_data.entrySet()) {
-                options_root.put(data.getKey().getName(), data.getValue());
-            }
+            msg_root.put("options", options_data);
         }
 
         // method
         HashMap<String, Object> child_message = null;
         for (Descriptors.MethodDescriptor sub_desc : msg_desc.getMethods()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -417,24 +406,21 @@ public class DataDstPbHelper {
         return null;
     }
 
-    static public void dumpOptionsIntoHashMap(LinkedList<Object> parent, Descriptors.FileDescriptor file_desc) {
+    static public void dumpOptionsIntoHashMap(LinkedList<Object> parent, Descriptors.FileDescriptor file_desc,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
         HashMap<String, Object> msg_root = new HashMap<String, Object>();
         boolean has_data = false;
 
-        Map<Descriptors.FieldDescriptor, Object> options_data = file_desc.getOptions().getAllFields();
-        if (!options_data.isEmpty()) {
-            HashMap<String, Object> options_root = new HashMap<String, Object>();
+        HashMap<String, Object> options_data = dumpMessageExtensions(file_desc.getOptions(),
+                file_desc.getOptions().getAllFields(), custom_extensions);
+        if (options_data != null && !options_data.isEmpty()) {
             has_data = true;
-            msg_root.put("options", options_root);
-
-            for (Map.Entry<Descriptors.FieldDescriptor, Object> data : options_data.entrySet()) {
-                options_root.put(data.getKey().getName(), data.getValue());
-            }
+            msg_root.put("options", options_data);
         }
 
         HashMap<String, Object> child_message = null;
         for (Descriptors.Descriptor sub_desc : file_desc.getMessageTypes()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -451,7 +437,7 @@ public class DataDstPbHelper {
         // enum_type
         child_message = null;
         for (Descriptors.EnumDescriptor sub_desc : file_desc.getEnumTypes()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -468,7 +454,7 @@ public class DataDstPbHelper {
         // service
         child_message = null;
         for (Descriptors.ServiceDescriptor sub_desc : file_desc.getServices()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -485,7 +471,7 @@ public class DataDstPbHelper {
         // extensions
         child_message = null;
         for (Descriptors.FieldDescriptor sub_desc : file_desc.getExtensions()) {
-            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc);
+            HashMap<String, Object> subset = dumpOptionsIntoHashMap(sub_desc, custom_extensions);
             if (subset == null) {
                 continue;
             }
@@ -506,5 +492,336 @@ public class DataDstPbHelper {
 
             parent.add(msg_root);
         }
+    }
+
+    private static HashMap<String, Object> convertMessageIntoInnerMap(Message msg) {
+        HashMap<String, Object> ret = new HashMap<String, Object>();
+        if (msg == null) {
+            return ret;
+        }
+
+        for (Map.Entry<Descriptors.FieldDescriptor, Object> kv : msg.getAllFields().entrySet()) {
+            switch (kv.getKey().getJavaType()) {
+            case INT:
+            case LONG:
+            case FLOAT:
+            case DOUBLE:
+            case BOOLEAN:
+            case STRING:
+            case BYTE_STRING: {
+                ret.put(kv.getKey().getName(), kv.getValue());
+                break;
+            }
+            case ENUM: {
+                if (kv.getValue() instanceof List<?>) {
+                    ArrayList<Integer> res = new ArrayList<Integer>();
+                    res.ensureCapacity(((List<?>) kv.getValue()).size());
+
+                    for (Object enum_value_desc : (List<?>) kv.getValue()) {
+                        if (enum_value_desc instanceof EnumValueDescriptor) {
+                            res.add(((EnumValueDescriptor) enum_value_desc).getNumber());
+                        }
+                    }
+                    ret.put(kv.getKey().getName(), res);
+                } else if (kv.getValue() instanceof EnumValueDescriptor) {
+                    ret.put(kv.getKey().getName(), ((EnumValueDescriptor) kv.getValue()).getNumber());
+                }
+                break;
+            }
+            case MESSAGE: {
+                if (kv.getValue() instanceof List<?>) {
+                    ArrayList<HashMap<String, Object>> res = new ArrayList<HashMap<String, Object>>();
+                    res.ensureCapacity(((List<?>) kv.getValue()).size());
+
+                    for (Object submsg : (List<?>) kv.getValue()) {
+                        if (submsg instanceof Message) {
+                            res.add(convertMessageIntoInnerMap((Message) submsg));
+                        }
+                    }
+                    ret.put(kv.getKey().getName(), res);
+                } else if (kv.getValue() instanceof Message) {
+                    ret.put(kv.getKey().getName(), convertMessageIntoInnerMap((Message) kv.getValue()));
+                }
+                break;
+            }
+            default:
+                break;
+            }
+        }
+
+        return ret;
+    }
+
+    private static Object pickMessageField(com.google.protobuf.UnknownFieldSet.Field field_data,
+            FieldDescriptor field_desc, Message defaultInstance)
+            throws InvalidProtocolBufferException, UnsupportedEncodingException {
+        if (null == field_desc || null == field_data) {
+            return null;
+        }
+
+        switch (field_desc.getType()) {
+        case DOUBLE: {
+            List<Long> raw_data = field_data.getFixed64List();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                ArrayList<Double> ret = new ArrayList<Double>();
+                ret.ensureCapacity(raw_data.size());
+                for (Long v : raw_data) {
+                    ret.add(Double.longBitsToDouble(v));
+                }
+                return ret;
+            } else {
+                return Double.longBitsToDouble(raw_data.get(0));
+            }
+        }
+        case FLOAT: {
+            List<Integer> raw_data = field_data.getFixed32List();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                ArrayList<Float> ret = new ArrayList<Float>();
+                ret.ensureCapacity(raw_data.size());
+                for (Integer v : raw_data) {
+                    ret.add(Float.intBitsToFloat(v));
+                }
+                return ret;
+            } else {
+                return Float.intBitsToFloat(raw_data.get(0));
+            }
+        }
+        case INT64:
+        case UINT64:
+        case INT32:
+        case UINT32: {
+            List<Long> raw_data = field_data.getVarintList();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                return raw_data;
+            } else {
+                return raw_data.get(0);
+            }
+        }
+
+        case SINT64: {
+            List<Long> raw_data = field_data.getVarintList();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                ArrayList<Long> ret = new ArrayList<Long>();
+                ret.ensureCapacity(raw_data.size());
+                for (int i = 0; i < raw_data.size(); ++i) {
+                    ret.add((raw_data.get(i) << 63) ^ (raw_data.get(i) >> 63));
+                }
+                return raw_data;
+            } else {
+                return (raw_data.get(0) << 63) ^ (raw_data.get(0) >> 63);
+            }
+        }
+        case SINT32: {
+            List<Long> raw_data = field_data.getVarintList();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                ArrayList<Integer> ret = new ArrayList<Integer>();
+                ret.ensureCapacity(raw_data.size());
+                for (int i = 0; i < raw_data.size(); ++i) {
+                    ret.add((raw_data.get(i).intValue() << 31) ^ (raw_data.get(i).intValue() >> 31));
+                }
+                return raw_data;
+            } else {
+                return (raw_data.get(0).intValue() << 31) ^ (raw_data.get(0).intValue() >> 31);
+            }
+        }
+
+        case ENUM: {
+            List<Long> raw_data = field_data.getVarintList();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                ArrayList<Integer> ret = new ArrayList<Integer>();
+                ret.ensureCapacity(raw_data.size());
+                for (int i = 0; i < raw_data.size(); ++i) {
+                    EnumValueDescriptor enum_value_desc = field_desc.getEnumType()
+                            .findValueByNumber(raw_data.get(0).intValue());
+                    if (null == enum_value_desc) {
+                        continue;
+                    }
+                    ret.add(enum_value_desc.getNumber());
+                }
+                return ret;
+            } else {
+                EnumValueDescriptor enum_value_desc = field_desc.getEnumType()
+                        .findValueByNumber(raw_data.get(0).intValue());
+                if (null == enum_value_desc) {
+                    return null;
+                }
+                return enum_value_desc.getNumber();
+            }
+        }
+        case BOOL: {
+            List<Long> raw_data = field_data.getVarintList();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                ArrayList<Boolean> ret = new ArrayList<Boolean>();
+                ret.ensureCapacity(raw_data.size());
+                for (int i = 0; i < raw_data.size(); ++i) {
+                    ret.add(raw_data.get(i) != 0);
+                }
+                return ret;
+            } else {
+                return raw_data.get(0) != 0;
+            }
+        }
+        case SFIXED64:
+        case FIXED64: {
+            List<Long> raw_data = field_data.getFixed64List();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                return raw_data;
+            } else {
+                return raw_data.get(0);
+            }
+        }
+        case SFIXED32:
+        case FIXED32: {
+            List<Integer> raw_data = field_data.getFixed32List();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                return raw_data;
+            } else {
+                return raw_data.get(0);
+            }
+        }
+        case STRING: {
+            List<ByteString> raw_data = field_data.getLengthDelimitedList();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                ArrayList<String> ret = new ArrayList<String>();
+                ret.ensureCapacity(raw_data.size());
+                for (int i = 0; i < raw_data.size(); ++i) {
+                    ret.add(new String(raw_data.get(i).toByteArray(), "UTF-8"));
+                }
+                return ret;
+            } else {
+                return new String(raw_data.get(0).toByteArray(), "UTF-8");
+            }
+        }
+        case GROUP: // ignore & skip
+            return null;
+        case MESSAGE: {
+            List<ByteString> raw_data = field_data.getLengthDelimitedList();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                ArrayList<HashMap<String, Object>> ret = new ArrayList<HashMap<String, Object>>();
+                ret.ensureCapacity(raw_data.size());
+                for (int i = 0; i < raw_data.size(); ++i) {
+                    Message.Builder builder = defaultInstance.newBuilderForType();
+                    ret.add(convertMessageIntoInnerMap(builder.mergeFrom(raw_data.get(i)).build()));
+                }
+                return ret;
+            } else {
+                Message.Builder builder = defaultInstance.newBuilderForType();
+                return convertMessageIntoInnerMap(builder.mergeFrom(raw_data.get(0)).build());
+            }
+        }
+        case BYTES: {
+            List<ByteString> raw_data = field_data.getLengthDelimitedList();
+            if (raw_data == null || raw_data.isEmpty()) {
+                return null;
+            }
+
+            if (field_desc.isRepeated()) {
+                return raw_data;
+            } else {
+                return raw_data.get(0);
+            }
+        }
+        default:
+            return null;
+        }
+    }
+
+    private static <T extends com.google.protobuf.GeneratedMessageV3.ExtendableMessage<T>> HashMap<String, Object> dumpMessageExtensions(
+            T msg, Map<Descriptors.FieldDescriptor, Object> options,
+            com.google.protobuf.ExtensionRegistry custom_extensions) {
+        HashMap<String, Object> ret = null;
+        if (null == custom_extensions) {
+            return ret;
+        }
+
+        if (options != null && !options.isEmpty()) {
+            ret = new HashMap<String, Object>();
+            for (Map.Entry<Descriptors.FieldDescriptor, Object> kv : options.entrySet()) {
+                ret.put(kv.getKey().getName(), kv.getValue());
+            }
+        }
+
+        Map<Integer, com.google.protobuf.UnknownFieldSet.Field> all_unknown_fields = msg.getUnknownFields().asMap();
+        if (all_unknown_fields.isEmpty()) {
+            return ret;
+        }
+
+        Set<ExtensionInfo> all_ext_types = custom_extensions
+                .getAllImmutableExtensionsByExtendedType(msg.getDescriptorForType().getFullName());
+
+        if (all_ext_types.isEmpty()) {
+            return ret;
+        }
+
+        for (ExtensionInfo ext_type : all_ext_types) {
+            com.google.protobuf.UnknownFieldSet.Field field_data = all_unknown_fields
+                    .getOrDefault(ext_type.descriptor.getNumber(), null);
+            if (field_data == null) {
+                continue;
+            }
+
+            if (ret != null && ret.containsKey(ext_type.descriptor.getName())) {
+                continue;
+            }
+
+            try {
+                Object val;
+                val = pickMessageField(field_data, ext_type.descriptor, ext_type.defaultInstance);
+                if (null != val) {
+                    if (ret == null) {
+                        ret = new HashMap<String, Object>();
+                    }
+                    ret.put(ext_type.descriptor.getName(), val);
+                }
+            } catch (InvalidProtocolBufferException | UnsupportedEncodingException e) {
+                // Ignore error
+            }
+        }
+
+        return ret;
     }
 }