+++ src/org/xresloader/core/data/dst/DataDstUEJson.java
@@ -159,7 +159,18 @@ public class DataDstUEJson extends DataDstUEBase {
         // Hashmap
         if (data instanceof Map) {
             Map<String, Object> mp = (Map<String, Object>) data;
-            for (Map.Entry<String, Object> item : mp.entrySet()) {
+            ArrayList<Map.Entry<String, Object>> sorted_array = new ArrayList<Map.Entry<String, Object>>();
+            sorted_array.ensureCapacity(mp.size());
+            sorted_array.addAll(mp.entrySet());
+            sorted_array.sort((l, r) -> {
+                if (l.getValue() instanceof Integer && r.getValue() instanceof Integer) {
+                    return ((Integer) l.getValue()).compareTo((Integer) r.getValue());
+                }
+
+                return 0;
+            });
+
+            for (Map.Entry<String, Object> item : sorted_array) {
                 if (prefix.isEmpty()) {
                     writeConstData(jo, item.getValue(), String.format("%s", item.getKey()), valSeg);
                 } else {